Move logger to new log folder
