+++ src/biocircuits/__init__.py
@@ -1,4 +1,4 @@
 from . import datasets
 from . import arch
-from .util.logger import Logger
+from .log.logger import Logger
 from .util.online_trainer import OnlineTrainer
